Refactor loot tab into partial form
